+++ src/qt/syscoinunits.cpp
@@ -109,7 +109,7 @@ QString SyscoinUnits::format(int unit, const CAmount& nIn, const uint64_t &nAsse
     uint8_t nPrecision = 8;
     if(nAsset > 0 && GetAssetPrecision(GetBaseAssetID(nAsset), nPrecision)) {
         num_decimals = (int)nPrecision;
-        coin = (qint64)pow(10.0, num_decimals);
+        coin = (qint64)powf(10.0, num_decimals);
     } else {
         // Note: not using straight sprintf here because we do NOT want
         // localized number formatting.
